[dev]优化LuaPlugin及LuaEnvObj对象

+++ src/HelloOgre3D/sandbox/manager/ObjectManager.cpp
@@ -302,6 +302,6 @@ void ObjectManager::markNodeRemInSeconds(Ogre::SceneNode* pSceneNode, float seco
 {
 	if (pSceneNode == nullptr)
 		return;
-	m_remSceneNodes[pSceneNode] = seconds * 1000;
+	m_remSceneNodes[pSceneNode] = int(seconds * 1000);
 }
